requirements.txt
Flask
flask-cors
mysql-connector-python 